[ci skip] Make the upward merge for non-stable packaging more obvious
The notice to merge upward is easy to miss when pushing the other
branches and the RC tag. Call attention to it with code blocks, and move
the note on resolving merge conflicts from an upward merge into a note.

Perhaps the upward merge merits its own step and thus a renumbering?