mail: fix exit code handling of sendmail cmd
Prior to this commit the return code of the pclose function was assumed
to be the exit code of the process. However, the returned value as
specified in wait(2) is a bit packed integer and must be interpreted
with the provided macros. After this commit we use the macros to obtain
the status code and also add some logging on failure.

For WIN32 we only check if the return value is non-zero, which should
solve, #43327