Remove zend_jit_vm_kind
JIT used to not have a compile-time dependency on VM kind, such that a single
build of opcache could work with VM kinds at runtime. This has been broken
over time and would be difficult to restore. Additionally, as opcache is now
built-in, this would not be useful anymore.

Remove the zend_jit_vm_kind variable.