Fix leak with EXT_STMT when pipe result is unused
This example would leak:

    function x() {
        return range(0, 10);
    }
    function test($a, $b) {
        $a |> $b;
    }
    test(42, x(...));

because we were missing a FREE in test(). This requires two fixes, namely one in
zend_do_free() to skip over the EXT_STMT usage, and another in
zend_optimize_block() to avoid removing the FREE opcode by removing the result
def of the declaring opcode.