uri: Remove useless non-NULL assertions in uri_parser_rfc3986.c
The value of `uriparser_uri` is dereferenced *immediately* after the assertion,
making it redundant for both the human reader and the compiler. In case of the
normalized_uri, the pointer is trivially non-NULL, since it is referring to a
struct member at not-the-first position.

Generally speaking, unless something is checked for `NULL` one can assume it to
never be `NULL`.