uri: Stop touching the `lxb_url_parser_t.idna` field
Looking at the Lexbor implementation, the `lxb_url_parser_t.idna` field is
private and must not be touched from the outside. Lexbor expects to be able to
manage it by itself when destroying a parser object.

Fix the issue by putting the `lxb_unicode_idna_t` into a thread-local variable
that we own. This also avoids one level of dynamic allocation. The same is done
for the mraw.