uri: Remove the `uri_parser` parameter from `php_uri_instantiate_uri()`
Similarly to the previous change to `uri_unserialize()`, the `->parser()` must
always match the object for the freeing to be safe.

Given that we expect to successfully parse URIs, we can eagerly initialize the
resulting URI object when using the `::parse()` methods and destruct it again
when parsing fails and `null` is returned instead. The destructor has
previously been made safe for this case by initializing the `->uri` field to
`NULL`.

The `base_url_object` must also match the object that is currently being
constructed. Verify this using assertions matching the `->ce` and the
`->parser`.