Remove include "sanity check" to get better error (#19650)
* Remove include "sanity check" to get a better error message

Instead of rejecting directories / non-regular files early with
a generic error, we should just accept them and error later when a
read is attempted. This is more general and will generate a better
error message on Linux. On Windows, the error remains the same as
before.

* Update error message to include include_path

fix format for include path

---------

Co-authored-by: Joe Watkins <krakjoe@php.net>