Fix GH-19653: Closure named argument unpacking between temporary closures can cause a crash
Due to closures, the `fbc` address isn't unique if the memory address is reused.
So for user closures we need to distinguish using a unique key,
while internal functions can't disappear and therefore can use the `fbc` address
as unique key.

As for performance:
This adds a load+compare for the function type, although we already load
and compare the function type down below, so that impact on performance
should be small. It also adds a load on the opcodes array, but we access
fields from the same cacheline too so that should also have only a small
impact.