zend_vm_gen: Track line numbers on a per-file basis (#19799)
php/php-src#19789 fixed the line number references for `zend_vm_def.h`, when
generating with line number information some of them are also specific to
`zend_vm_execute.h` and thus should reference that file instead with the
correct line numbers.

The line number tracking was broken, because it was tracked in a single global
variable, instead of being tracked on a per-file basis. Fix this by making the
line numbers an array indexed by the resource ID and consistently using the
`out()` functions to write into the files.

Co-authored-by: Arnaud Le Blanc <365207+arnaud-lb@users.noreply.github.com>