ODBC fetch refactoring (#19848)
* Merge odbc_fetch_into into odbc_fetch_hash

Now that we can assume fetch_hash exists, there's a lot of redundancy in
these functions. Merge their implementations, and smooth over the
differences in how they handle returning their result set as an array.

* Convert php_odbc_fetch_hash to ZPP

* Use SQLFetchScroll instead of SQLExtendedFetch

Fixes GH-19522

* Convert result type constants to an enum

* Implement odbc_fetch_row in terms of php_odbc_fetch_hash

These are also doing extremely similar jobs, but with slightly different
behaviours for the return value (in this case, none, as it's tended to
be used with odbc_result). Unify this too.

The $row value deprecation for 0/-1 is only handled for odbc_fetch_row;
it's too late to do so for PHP 8.5. Should probably unify it for PHP 8.6.

* Rename php_odbc_fetch_hash to remove _hash

Since this is a much more shared fetch function now.