Fix flaky logging test
As documented in [RunningTests], tests have to be written to be
independent of any php.ini file.

Without specifying that no errors should be logged, the php cli server
invocation with the invalid address provokes additional diagnostic PHP
Warning messages on the diagnostic channel which invalidates the EXPECTF
 section.

This renders the test-case flaky and will make it fail unless
`log_errors` is `0`.

Fix is to set the directive on the command line.

[RunningTests]: docs/source/miscellaneous/running-tests.rst