Support specifying an exact date/time range for validity when signing a CSR
This changes the `days` parameter of `openssl_csr_sign()` to a
`validity` parameter, which can be either an integer specifying the
number of days the certificate is to be valid for (compatible with
current usage), or it can be an array of two integer or string values,
representing the notBefore and notAfter times to use for the
certificate.  If they are integers or numeric strings, they are to be a
time_t value.  If they are non-numeric strings, they are to be an ASN.1
timestamp (YYMMDDHHMMSSZ or YYYYMMDDHHMMSSZ).